requirements.txt
pyautogen
python-dotenv
openai
streamlit
pytest